Split grpc-related files into multiple directories
