+++ pinecone/grpc/protos/vector_column_service_pb2_grpc.py
